app/requirements.txt
argparse
datasets
openai
pinecone-client
python-dotenv
tqdm
uuid